requirements.txt
aio-pika==9.5.7
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiormq==6.9.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
attrs==25.4.0
camelot-py==1.0.9
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
distro==1.9.0
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.120.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
honcho==2.0.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
minio==7.2.18
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
openai==2.6.1
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pamqp==3.3.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
playwright==1.55.0
propcache==0.4.1
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.26.5
pypdf==5.9.0
pypdfium2==4.30.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rich==14.2.0
rpds-py==0.28.0
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.10.16
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
